# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fake Store API",
        "description": "This is the API documentation for my Laravel project.",
        "contact": {
            "email": "your-email@example.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/staffs": {
            "get": {
                "tags": [
                    "Staff"
                ],
                "summary": "Get paginated list of staff",
                "operationId": "69e3011ec7251b70145144de8e58aac0",
                "parameters": [
                    {
                        "name": "_pageLimit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "_pageSize",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful pagination response"
                    }
                }
            }
        },
        "/api/staff/{id}": {
            "get": {
                "tags": [
                    "Staff"
                ],
                "summary": "Get staff by ID",
                "operationId": "4159e1238559ccb72045e777a1f615f4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Staff found"
                    },
                    "404": {
                        "description": "Staff not found"
                    }
                }
            }
        },
        "/api/staff/create": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Create a new staff member",
                "description": "Creates a new staff record with validated fields.",
                "operationId": "a51efdba6969031b69ea27b59b9f47d5",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "position_id",
                                    "name",
                                    "dob",
                                    "pob",
                                    "address",
                                    "phone_number",
                                    "nation_id_card"
                                ],
                                "properties": {
                                    "position_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "gender": {
                                        "type": "string",
                                        "example": "Male",
                                        "nullable": true
                                    },
                                    "dob": {
                                        "type": "string",
                                        "example": "1990-01-01"
                                    },
                                    "pob": {
                                        "type": "string",
                                        "example": "Phnom Penh"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Street 123, Phnom Penh"
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "012345678"
                                    },
                                    "nation_id_card": {
                                        "type": "string",
                                        "example": "N123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Staff created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "new_staff": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "position_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "John Doe"
                                                },
                                                "gender": {
                                                    "type": "string",
                                                    "example": "Male"
                                                },
                                                "dob": {
                                                    "type": "string",
                                                    "example": "1990-01-01"
                                                },
                                                "pob": {
                                                    "type": "string",
                                                    "example": "Phnom Penh"
                                                },
                                                "address": {
                                                    "type": "string",
                                                    "example": "Street 123"
                                                },
                                                "phone_number": {
                                                    "type": "string",
                                                    "example": "012345678"
                                                },
                                                "nation_id_card": {
                                                    "type": "string",
                                                    "example": "N123456789"
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "example": "2025-07-11T12:00:00.000000Z"
                                                },
                                                "updated_at": {
                                                    "type": "string",
                                                    "example": "2025-07-11T12:00:00.000000Z"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 200
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "position_id": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The selected position ID does not exist."
                                            }
                                        },
                                        "name": {
                                            "type": "array",
                                            "items": {
                                                "type": "string",
                                                "example": "The name field is required."
                                            }
                                        },
                                        "status_code": {
                                            "type": "integer",
                                            "example": 422
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/api/staff/delete/force": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Force delete staff",
                "operationId": "3a8ee22f795e7af7c5b35932a52a0f4c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Staff permanently deleted"
                    },
                    "404": {
                        "description": "Staff not found in trash"
                    }
                }
            }
        },
        "/api/staff/delete/soft": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Soft delete staff",
                "operationId": "8f7c9478ecbf6207958c867211da3768",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Staff soft deleted"
                    },
                    "404": {
                        "description": "Staff not found"
                    }
                }
            }
        },
        "/api/staff/restore/{id}": {
            "post": {
                "tags": [
                    "Staff"
                ],
                "summary": "Restore soft deleted staff",
                "operationId": "b6e0fb55c6876d14e22552f9974ecd09",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Staff restored successfully"
                    },
                    "404": {
                        "description": "Staff not found in trash"
                    }
                }
            }
        }
    },
...
}